py_snowflake_poc/py_snowflake/requirements.txt
asn1crypto==1.3.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.11.17
botocore==1.14.17
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
ijson==2.6.1
jmespath==0.9.5
oscrypto==1.2.0
pycparser==2.19
pycryptodomex==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
snowflake-connector-python==2.2.1
urllib3==1.25.8
